Add spring security authentication for journal entry controller

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -30,8 +30,8 @@ public class JournalEntryService {
             user.getJournalEntries().add(saved);
             userService.saveUser(user);
         } catch (Exception e) {
-           System.out.println(e);
-           throw new RuntimeException("An error has occured while saving the entry");
+            System.out.println(e);
+            throw new RuntimeException("An error has occured while saving the entry");
         }
     }
 
@@ -47,11 +47,21 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName) {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveUser(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed;
+        try{
+User user = userService.findByUserName(userName);
+        removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        if (removed) {
+            userService.saveUser(user);
+            journalEntryRepository.deleteById(id);
+        }
+        }catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting the entry.",e);
+        }
+        return removed;
+
     }
-    // add service methods here
 }
